% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.calculateProp}
\alias{.calculateProp}
\title{Cell neighbourhood composition}
\usage{
.calculateProp(arr)
}
\arguments{
\item{arr}{a vector of initial subcluster labels of each cell in the
neighbourhood.}
}
\value{
a table of initial subcluster proportions in a neighbourhood.
}
\description{
A function to calculate the cell neighbourhood composition of initial cluster
labels.
}
\keyword{internal}
