% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourName.R
\name{.cellName}
\alias{.cellName}
\title{Neighbour cell naming}
\usage{
.cellName(cell, Clust)
}
\arguments{
\item{cell}{a vector of neighbourhood cell indices. The cell indices indicate
the row number of cells in sample metadata.}

\item{Clust}{a data frame of initial cluster assignments of all cells in the
sample.}
}
\value{
a data frame of cell IDs of neighbourhood cells.
}
\description{
A function to fetch cell IDs.
}
\keyword{internal}
