% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.clustNum}
\alias{.clustNum}
\title{Neighbour cell initial subcluster label}
\usage{
.clustNum(cell, subClust)
}
\arguments{
\item{cell}{a vector of neighbourhood cell indices. The cell indices indicate
the row number of cells in sample metadata.}

\item{subClust}{a data frame of initial subcluster labels of each cell in
the sample.}
}
\value{
a data frame of initial subcluster labels of neighbourhood
cells.
}
\description{
A function to fetch initial subcluster label of cell.
}
\keyword{internal}
