% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.exp_kernel}
\alias{.exp_kernel}
\title{Exponential distribution weights}
\usage{
.exp_kernel(ed, NN, rate)
}
\arguments{
\item{ed}{a numeric vector of entropy values of all cell neighbourhoods.}

\item{NN}{an integer for the number of neighbourhood cells including the
index cell.}

\item{rate}{a numeric value for rate of exponential distribution.}
}
\value{
a matrix where the columns contain weights associated with the
entropy values.
}
\description{
A function to generate weights from an exponential distribution.
}
\keyword{internal}
