% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.gauss_kernel}
\alias{.gauss_kernel}
\title{Gaussian distribution weights}
\usage{
.gauss_kernel(ed, NN, sd)
}
\arguments{
\item{ed}{a numeric vector of entropy values of all cell neighbourhoods.}

\item{NN}{an integer for the number of neighbourhood cells including the
index cell.}

\item{sd}{a numeric value for standard deviation of Gaussian distribution.}
}
\value{
a matrix where the columns contain weights associated with the
entropy values.
}
\description{
A function to generate weights from a Gaussian distribution.
}
\keyword{internal}
