% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{.getColours}
\alias{.getColours}
\title{Plot colours}
\usage{
.getColours()
}
\value{
a vector of colours.
}
\description{
A vector of colours used in clustSIGNAL spatial plots by default.
}
\keyword{internal}
