% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{.metrics}
\alias{.metrics}
\title{Clustering metrics}
\usage{
.metrics(spe, samples, celltypes, cells)
}
\arguments{
\item{spe}{spatialExperiment object with logcounts, PCA, initial non-spatial clustering, entropy, and smoothed gene expression outputs included.}

\item{samples}{samples a character vector of sample names to which the cells belong. Length of vector must be equal to the number of cells in spatialExperiment object (i.e. the number of rows in colData(spe)).}

\item{celltypes}{a character vector of cell type annotation of each cell. Length of vector must be equal to the number of cells in spatialExperiment object (i.e. the number of rows in colData(spe)).}

\item{cells}{a character vector of cell IDs of each cell. Length of vector must be equal to the number of cells in spatialExperiment object (i.e. the number of rows in colData(spe)).}
}
\value{
a list containing two items:

1. metrics_out, a table of clustering metrics ARI, NMI, and ASW, and domainness statistics min, max, and mean entropy values calculated for each sample.

2. spe_out, spatialExperiment object with annotated cell type and post-smoothing cluster silhouette widths of each cell added as metadata.
}
\description{
A function to measure clustering performance using ARI, NMI, and silhouette width, and to provide entropy statistics of each sample in the dataset.
}
