% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plots}
\alias{.plots}
\title{Visualisation plots}
\usage{
.plots(spe, samples, celltypes, plotType)
}
\arguments{
\item{spe}{SpatialExperiment object. For plotType = domain, the object should contain entropy values, for plotType = spatial, the object should contain cluster outputs, and for plotType = heatmap, the object should contain silhouette widths and cluster outputs.}

\item{samples}{a character vector of sample names to which the cells belong. Length of vector must be equal to the number of cells in SpatialExperiment object (i.e. the number of rows in colData(spe)).}

\item{celltypes}{a character vector of cell type annotation of each cell. Length of vector must be equal to the number of cells in SpatialExperiment object (i.e. the number of rows in colData(spe)).}

\item{plotType}{a character parameter handled within the method, to decide what plots to generate.}
}
\value{
generates plots depending on the plotType parameter:

plotType = domain,  generates a histogram of entropy values and number of cell neighbourhoods and a spatial plot showing entropy of all cell neighbourhoods.

plotType = spatial, generates spatial plots of annotated cell types and clusters generated from smoothed data.

plotType = heatmap, generates a heat map showing cell number distributions of annotated cell types and post-smoothing clusters.
}
\description{
A function to visualise the outputs of clustSIGNAL, including plots for domainness spread and distribution, spatial plots of annotated cells and final clusters, and heatmap comparison of annotated cell types and final clusters.
}
