% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyMeasure.R
\name{entropyMeasure}
\alias{entropyMeasure}
\title{Heterogeneity measure}
\usage{
entropyMeasure(spe, regXclust)
}
\arguments{
\item{spe}{SpatialExperiment object with initial cluster and subcluster
labels.}

\item{regXclust}{a numeric matrix of cells by subclusters, where the values
are the proportion of initial subclusters in each cell's neighbourhood.}
}
\value{
SpatialExperiment object with entropy values associated with each
cell.
}
\description{
A function to measure the heterogeneity of a cell's neighbourhood in terms of
entropy. Generally, homogeneous neighbourhoods have low entropy and
heterogeneous neighbourhoods have high entropy.
}
\examples{
data(ClustSignal_example)

# requires matrix containing cluster proportions of each neighbourhood
# (regXclust), generated using the neighbourDetect() function
spe <- clustSIGNAL::entropyMeasure(spe, regXclust)
spe$entropy |> head()
}
