% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mEmbryo2}
\alias{mEmbryo2}
\alias{me_data}
\alias{me_expr}
\title{Mouse Embryo Data}
\format{
\code{me_expr} a gene expression matrix with normalised counts, where rows indicate
genes and columns indicate cells.
\code{me_data} a data frame of cell metadata including cell IDs, sample IDs,
cell type annotations, and x-y coordinates of cells.
}
\source{
Integration of spatial and single-cell transcriptomic data elucidates mouse
organogenesis, \emph{Nature Biotechnology}, 2022.
Webpage: \url{https://www.nature.com/articles/s41587-021-01006-2}
}
\usage{
data(mEmbryo2)
}
\description{
This dataset contains spatial transcriptomics data from 3 mouse embryos, with
351 genes and a total of 57536 cells. For vignettes, we subset the data by randomly
selecting 5000 cells from embryo 2, excluding cells that were annotated as 'low quality'.
}
\keyword{datasets}
