% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mHypothal}
\alias{mHypothal}
\alias{mh_data}
\alias{mh_expr}
\title{Mouse Hypothalamus Data as SpatialExperiment object}
\format{
\code{mh_expr} a gene expression matrix with normalised counts, where rows indicate
genes and columns indicate cells.
\code{mh_data} a data frame of cell metadata including cell IDs, sample IDs,
cell type annotations, and x-y coordinates of cells.
}
\source{
Molecular, Spatial and Functional Single-Cell Profiling of the
Hypothalamic Preoptic Region, \emph{Science}, 2018.
Webpage: \url{https://www.science.org/doi/10.1126/science.aau5324}
}
\usage{
load("mHypothal.RData")
}
\description{
This dataset contains spatial transcriptomics data from 181 mouse hypothalamus
samples embryos, 155 genes and a total of 1,027,080 cells. For running the
vignettes, we subset the data by selecting only 3 samples - Animal 1 Bregma -0.09
and Animal 7 Bregmas 0.16 and -0.09, removed all cells that were annotated
as 'ambiguous', and removed 20 genes that were assessed using a different technology.
After subsetting, we have 15,848 cells from 3 mouse brain samples and 135 genes.
}
\keyword{datasets}
