% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{nsClustering}
\alias{nsClustering}
\title{Non-spatial clustering}
\usage{
nsClustering(spe, reclust)
}
\arguments{
\item{spe}{SpatialExperiment object. For reclust = FALSE, the object should contain logcounts and PCA, but for reculst = TRUE, the object should contain smoothed gene expression.}

\item{reclust}{a logical parameter handled within the method.}
}
\value{
SpatialExperiment object containing 'putative cell type' group allotted to each cell (reclust = FALSE) or clusters generated from smoothed data (reclust = TRUE).
}
\description{
A function containing two steps used at different times in the clustSIGNAL workflow. An initial non-spatial clustering and sub-clustering step (reclust = FALSE) is used to generate groups of ‘putative cell types’, whereas a later non-spatial clustering step (reclust = TRUE) is used to cluster adaptively smoothed gene expression data.
}
