% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initClustering.R
\name{p1_clustering}
\alias{p1_clustering}
\title{Initial non-spatial clustering}
\usage{
p1_clustering(spe, dimRed, batch, batch_by, clustParams)
}
\arguments{
\item{spe}{SpatialExperiment object. For reclust = FALSE, the object should
contain logcounts and PCA, but for reculst = TRUE, the object should contain
smoothed gene expression.}

\item{dimRed}{a character indicating the name of the reduced dimensions to
use from the SpatialExperiment object (i.e., from reducedDimNames(spe)).
Default value is 'PCA'.}

\item{batch}{a logical parameter for whether or not to perform batch
correction. Default value is FALSE.}

\item{batch_by}{a character indicating name of colData(spe) column containing
the batch names.}

\item{clustParams}{a list of parameters for TwoStepParam clustering methods.
The clustering parameters are in the order - centers (centers) for clustering
with KmeansParam, centers (centers) for sub-clustering clusters with
KmeansParam, maximum iterations (iter.max) for clustering with KmeansParam,
k values (k) for clustering with NNGraphParam, and community detection method
(cluster.fun) to use with NNGraphParam.}
}
\value{
SpatialExperiment object containing 'initial cluster' groups
allotted to each cell.
}
\description{
A function to perform initial non-spatial clustering and sub-clustering of
normalised gene expression to generate 'initial clusters'.
}
\examples{
data(example)

spe <- clustSIGNAL::p1_clustering(spe, dimRed = "PCA", batch = FALSE,
                                 batch_by = "None",
                                 clustParams = list(0, 0, 30, 5, "louvain"))
head(spe$nsCluster)
head(spe$initCluster)

}
