% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustSIGNAL.R
\name{clustSIGNAL}
\alias{clustSIGNAL}
\title{clustSIGNAL}
\usage{
clustSIGNAL(
  spe,
  samples,
  cells,
  dimRed = "PCA",
  NN = 30,
  kernel = "G",
  spread = 0.05,
  sort = TRUE,
  threads = 1,
  outputs = "c",
  ...
)
}
\arguments{
\item{spe}{SpatialExperiment object with logcounts and PCA included.}

\item{samples}{a character vector of sample names to which the cells belong. Length of vector must be equal to the number of cells in SpatialExperiment object (i.e. the number of rows in colData(spe)).}

\item{cells}{a character vector of cell IDs of each cell. Length of vector must be equal to the number of cells in SpatialExperiment object (i.e. the number of rows in colData(spe)).}

\item{NN}{an integer for the number of neighbourhood cells the function should consider. The value must be greater than or equal to 1. Default value is 30.}

\item{kernel}{a character for type of distribution to be used. The two valid values are "G" or "E". G for Gaussian distribution, and E for exponential distribution. Default value is "G".}

\item{spread}{a numeric value for distribution spread, represented by standard deviation for Gaussian distribution and rate for exponential distribution. Default value is 0.05 for Gaussian distribution and 20 for exponential distribution.}

\item{sort}{a logical parameter for whether or not to sort the neighbourhood after region description. Default value is TRUE.}

\item{threads}{a numeric value for the number of CPU cores to be used for the analysis. Default value set to 4 cores.}

\item{outputs}{a character for the type of output to return to the user. "c" for data frame of cell IDs and their respective cluster numbers (default), "n" for dataframe of clusters plus neighbourhood matrix, "s" for  dataframe of clusters plus final spatialExperiment object, or "a" for all outputs.}

\item{...}{additional parameters for TwoStepParam clustering methods. Include parameters for k for number of nearest neighbours and cluster.fun for selecting community detection method. Default values k = 5, cluster.fun = "louvain".}
}
\value{
a list of outputs

1. clusters: a data frame of cell names and their cluster classification.

2. neighbours: a matrix of cell names and the names of their NN nearest neighbour cells.

3. spe_final: a SpatialExperiment object with initial 'putative cell type' groups, entropy values, smoothed gene expression, post-smoothing clusters, and silhouette widths included.
}
\description{
A clustering method for cell type classification of spatial transcriptomics data. The tool generates and uses an adaptively smoothed, spatially informed gene expression data for clustering.
}
\examples{
spe = load(mouseEmbryo2)
clustSIGNAL <- function (spe, spe$embryo, spe$uniqueIDs)


}
