% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourComposition.R
\name{.calculateProp}
\alias{.calculateProp}
\title{Cell neighbourhood composition}
\usage{
.calculateProp(arr)
}
\arguments{
\item{arr}{a vector of 'putative cell type' assignments of each cell in the
neighbourhood.}
}
\value{
a table of 'putative cell type' proportions in a neighbourhood.
}
\description{
A function to calculate the cell neighbourhood composition of 'putative' cell
types.
}
\keyword{internal}
