% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbourCluster.R
\name{.clustNum}
\alias{.clustNum}
\title{Neighbour cell 'putative cell type' assignment}
\usage{
.clustNum(cell, subClust)
}
\arguments{
\item{cell}{a vector of neighbourhood cell indices. The cell indices indicate the row number of cells in sample metadata.}

\item{subClust}{a data frame of 'putative cell type' assignments of all cells in the sample.}
}
\value{
a data frame of 'putative cell type' assignments of neighbourhood cells.
}
\description{
A function to fetch 'putative cell type' assignments of cell.
}
\keyword{internal}
