% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expKernel.R
\name{.exp_kernel}
\alias{.exp_kernel}
\title{Exponential distribution weights}
\usage{
.exp_kernel(ed, NN, rate)
}
\arguments{
\item{ed}{a numeric value for entropy of the neighbourhood.}

\item{NN}{an integer for the number of neighbourhood cells the function
should consider. The value must be greater than or equal to 1. Default value
is 30.}

\item{rate}{a numeric value for rate of exponential distribution.}
}
\value{
a matrix where each column contains weights related to specific
entropy values.
}
\description{
A function to generate weights from an exponential distribution.
}
