% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyMeasure.R
\name{entropyMeasure}
\alias{entropyMeasure}
\title{Domainness measure}
\usage{
entropyMeasure(spe, cells, regXclust, threads)
}
\arguments{
\item{spe}{SpatialExperiment object with logcounts, PCA, and 'putative cell type' groups included.}

\item{cells}{a character vector of cell IDs of each cell. Length of vector must be equal to the number of cells in spatialExperiment object (i.e. the number of rows in colData(spe)).}

\item{regXclust}{a list of vectors for each cell's neighbourhood composition indicated by the proportion of 'putative cell type' groups it contains.}

\item{threads}{a numeric value for the number of CPU cores to be used for the analysis.}
}
\value{
SpatialExperiment object including entropy values for each cell neighbourhood.
}
\description{
A function to measure the heterogeneity of a cell's neighbourhood in terms of entropy, such that homogeneous neighbourhoods have low entropy and heterogeneous neighbourhoods have high entropy.
}
\examples{
data(mouseEmbryo2)

# requires list of containing cluster proportions of each region (regXclust),
# generated using the neighbourDetect() function
spe <- entropyMeasure(spe, cells = "uniqueID", regXclust, threads = 1)
head(spe$entropy)

}
