% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyMeasure.R
\name{entropyMeasure}
\alias{entropyMeasure}
\title{Domainness measure}
\usage{
entropyMeasure(spe, regXclust, threads = 1)
}
\arguments{
\item{spe}{SpatialExperiment object with logcounts, PCA, and 'putative cell
type' groups included.}

\item{regXclust}{a list of vectors for each cell's neighbourhood composition
indicated by the proportion of 'putative cell type' groups it contains.}

\item{threads}{a numeric value for the number of CPU cores to be used for the
analysis.}
}
\value{
SpatialExperiment object including entropy values for each cell
neighbourhood.
}
\description{
A function to measure the heterogeneity of a cell's neighbourhood in terms of
entropy, such that homogeneous neighbourhoods have low entropy and
heterogeneous neighbourhoods have high entropy.
}
\examples{
data(ClustSignal_example)

# requires list containing cluster proportions of each region (regXclust),
# generated using the neighbourDetect() function
spe <- clustSIGNAL::entropyMeasure(spe, regXclust)
spe$entropy |> head()

}
