% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborDetect.R
\name{neighbourDetect}
\alias{neighbourDetect}
\title{Cell neighbourhood detection}
\usage{
neighbourDetect(spe, samples, NN = 30, cells, sort = TRUE)
}
\arguments{
\item{spe}{SpatialExperiment object with logcounts, PCA, and 'putative cell type' groups included.}

\item{samples}{a character vector of sample names to which the cells belong. Length of vector must be equal to the number of cells in spatialExperiment object (i.e. the number of rows in colData(spe)).}

\item{NN}{an integer for the number of neighbourhood cells the function should consider. The value must be greater than or equal to 1. Default value is 30.}

\item{cells}{a character vector of cell IDs of each cell. Length of vector must be equal to the number of cells in SpatialExperiment object (i.e. the number of rows in colData(spe)).}

\item{sort}{a logical parameter for whether or not to sort the neighbourhood after region description. Default value is TRUE.}
}
\value{
a list containing two items:
\enumerate{
\item nnCells, a character matrix of NN nearest neighbours - rows are cells and columns are their nearest neighbours ranged from closest to farthest neighbour. For sort = TRUE, the neighbours belonging to the same 'putative cell type' group as the cell are moved closer to it.
\item regXclust, a list of vectors for each cell's neighbourhood composition indicated by the proportion of 'putative cell type' groups it contains.
}
}
\description{
A function to identify the neighbourhood of each cell. If sort = TRUE, the neighbourhoods are also sorted such that cells belonging to the same group as the central cell are arranged closer to it.
}
\examples{
data(example)

out_list <- neighbourDetect(spe, samples = "sample_id", NN = 30, cells = "uniqueID", sort = TRUE)
names(out_list)

}
