% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalClustering.R
\name{p2_clustering}
\alias{p2_clustering}
\title{Final non-spatial clustering}
\usage{
p2_clustering(spe, batch, batch_by, clustParams)
}
\arguments{
\item{spe}{SpatialExperiment object. For reclust = FALSE, the object should
contain logcounts and PCA, but for reculst = TRUE, the object should contain
smoothed gene expression.}

\item{batch}{a logical parameter for whether or not to perform batch
correction. Default value is FALSE.}

\item{batch_by}{a character indicating name of colData(spe) column containing
the batch names.}

\item{clustParams}{a list of parameters for TwoStepParam clustering methods.
The clustering parameters are in the order - centers (centers) for clustering
with KmeansParam, centers (centers) for sub-clustering clusters with
KmeansParam, maximum iterations (iter.max) for clustering with KmeansParam,
k values (k) for clustering with NNGraphParam, and community detection method
(cluster.fun) to use with NNGraphParam.}
}
\value{
SpatialExperiment object containing clusters generated from smoothed
data.
}
\description{
A function containing to cluster adaptively smoothed gene expression data.
}
\examples{
data(example)

# For non-spatial clustering of normalised counts
spe <- clustSIGNAL::p2_clustering(spe, batch = FALSE, batch_by = "None",
                                 clustParams = list(0, 0, 30, 5, "louvain"))
head(spe$clustSIGNAL)

}
